:recycle:  refactor: useLottie의 type을 확장하여 사용하는 LottieLoader가 되도록 수정
